testing video file uploaded
Video file of a walmart parking lot taken by a drone. The actual footage is too shaky since it's taken from a drone. The parking spots shift and move as the drone gets pushed by the wind, and since the parking spots are determined by pixels, this makes it less accurate.